api/requirements.txt
pandas
numpy
unidecode
scikit-learn
sqlalchemy
plotly
psycopg2
uvicorn
fastapi
datetime
python-dotenv